Support multiple signers
To better support multiple/external (i.e HSM) keys, it would be ideal to
support multiple signers.

This is done by collapsing the keysource interface and passed signer in to a
new interface, SignerSource. This has a new method to return a signer to be
used, which is called for each signing call.

To handle the simple single key case, a basic imlementation of this is provided